fix: performance issues from unecessary obstacle assignments to lanelets
Before the fix, the `EnoughSurroundingVehiclesFilter` would assign *all*
obstacles at *all* time steps to their corresponding lanelets.
But the assignment at all time steps is unecessary, because:
1. We do not consider the whole scenario, but only the time frame of the ego maneuver
2. We can stop with the assignment, once we find one time step,
 at which there are enough surrounding vehicles

With this fix, the obstacles will be assigned on demand.
This has no influence of on the resulting scenarios, because the
assignment happens *before*, the resulting scenarios are created.